Revert "14.7.2022 Review code"
This reverts commit e577d9fbf588ad93b566982797403cc63f8f7fbc.

+++ week1/Booking_room/backend/BookingRoom/src/main/java/com/backend/bean/Room.java
@@ -17,10 +17,10 @@ public class Room {
     private String roomID;
 
     @Column(name = "capacity")
-    private Integer capacity;
+    private int capacity;
 
     @Column(name = "price")
-    private Integer price;
+    private double price;
 
     @Column(name = "image")
     private String image;
